+++ packages/apps/kadena-docs/src/components/Layout/Landing.tsx
@@ -1,6 +1,6 @@
 import { Content } from './components';
-import { ILayout } from './types';
 
+import { ILayout } from '@/types/Layout';
 import React, { FC } from 'react';
 
 export const Landing: FC<ILayout> = ({ children }) => {
