--- packages/apps/kadena-docs/src/components/Layout/types.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react';
-
-export interface ILayout {
-  children?: ReactNode;
-}
